// src/app/personality/profile/models/pdb-category.enum.ts
export enum PdbCategory {
  None = 0,
  PopCulture = 1,
  Television = 2,
  Movies = 3,
  Sports = 5,
  Cartoons = 7,
  AnimeManga = 8,
  Comics = 9,
  Noteworthy = 10,
  Gaming = 11,
  Literature = 12,
  Theatre = 13,
  Musician = 14,
  Internet = 15,
  TheArts = 16,
  Business = 17,
  Religion = 18,
  Science = 21,
  Historical = 22,
  WebComics = 26,
  Superheroes = 27,
  Philosophy = 28,
  Kpop = 29,
  Traits = 30,
  PlotsArchetypes = 31,
  Concepts = 32,
  Music = 33,
  Franchises = 34,
  Culture = 35,
  Theories = 36,
  PollsIfYou = 37,
  YourExperience = 38,
  TypeComboYourType = 39,
  AskPdb = 40,
  PdbCommunity = 41,
  Nature = 42,
  Technology = 43
}